Show Movie List with favorite feature (#6)
* :recycle: Setup Hilt Dependency Injection

* Add Retrofit and OkHttp dependencies. (#3)

* :hammer_and_wrench: Add OkHttp and Retrofit dependencies

* :hammer_and_wrench: Add Retrofit and OkHttp dependencies. Sets up the initial network layer within the data module.

* :paintbrush: Setup Material 3 color and typography schemes (#4)

* Feature/movie list (#5)

* :tada: Setup Data and Domain models for movie list

* :sparkles: Implement movie repository

Adds the `MovieRepository` interface in the domain layer and its implementation `MovieRepositoryImpl` in the data layer. This also includes the Dagger Hilt module for providing `MovieRepository`.

* :recycle: Implement GetNowPlayingMovies use case

* :art: Implement basic MovieList UI layout

* :building_construction: add movie list screen

Adds the initial structure for the movie list screen, including the top bar, search bar, and placeholder movie cards. Also includes styling updates to the `MovieCardInfo` component for better padding.

* :lipstick: Make movie list scrollable

Adds a `verticalScroll` modifier to the main Column to enable scrolling when the content exceeds the screen size.

* :building_construction: Add MovieList feature implementation

* :art: Implement Movie List Search Functionality

* :art: Implement Room database for favorite movies

* :building_construction: Implement Coil for image loading and Movie Poster component